chore: major appointment system refactoring
