+++ src/main/java/seedu/address/model/ModelManager.java
@@ -13,8 +13,8 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.person.Person;
-import seedu.address.model.reminder.Appointment;
-import seedu.address.model.reminder.AppointmentList;
+import seedu.address.model.appointment.Appointment;
+import seedu.address.model.appointment.AppointmentList;
 
 /**
  * Represents the in-memory model of the address book data.
