+++ src/main/java/seedu/address/model/UserPrefs.java
@@ -11,7 +11,8 @@ import java.util.Objects;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.reminder.Appointment;
+import seedu.address.model.appointment.Appointment;
+import seedu.address.model.appointment.AppointmentTable;
 
 /**
  * Represents User's preferences.
