+++ src/main/java/seedu/address/model/appointment/exceptions/AppointmentNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.reminder.exceptions;
+package seedu.address.model.appointment.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified person.
