+++ src/main/java/seedu/address/model/appointment/exceptions/DuplicateAppointmentException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.reminder.exceptions;
+package seedu.address.model.appointment.exceptions;
 
 /**
  * Signals that the operation will result in duplicate Persons (Persons are considered duplicates if they have the same
