+++ src/main/java/seedu/address/ui/AppointmentListPanel.java
@@ -8,7 +8,7 @@ import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 import javafx.scene.layout.Region;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.model.reminder.Appointment;
+import seedu.address.model.appointment.Appointment;
 
 /**
  * Panel containing the list of persons.
