+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -21,7 +21,7 @@ import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.person.Person;
-import seedu.address.model.reminder.Appointment;
+import seedu.address.model.appointment.Appointment;
 import seedu.address.testutil.PersonBuilder;
 
 public class AddCommandTest {
