+++ src/test/java/seedu/address/logic/commands/AliasCommandTest.java
@@ -18,7 +18,7 @@ import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Person;
-import seedu.address.model.reminder.Appointment;
+import seedu.address.model.appointment.Appointment;
 
 public class AliasCommandTest {
     @Test
